+++ toolium/behave/environment.py
@@ -126,75 +126,65 @@ def before_scenario(context, scenario):
     # Read @no_driver tag
     no_driver = 'no_driver' in scenario.tags or 'no_driver' in scenario.feature.tags
 
-    bdd_common_before_scenario(context, scenario, no_driver)
+    # Initialize and connect driver wrapper
+    start_driver(context, no_driver)
+
+    # Add assert screenshot methods with scenario configuration
+    add_assert_screenshot_methods(context, scenario)
+
+    # Configure Jira properties
+    save_jira_conf()
+
+    context.logger.info("Running new scenario: %s", scenario.name)
 
     # Behave dynamic environment
     context.dyn_env.execute_before_scenario_steps(context)
 
 
-def bdd_common_before_scenario(context_or_world, scenario, no_driver=False):
-    """Common scenario initialization in behave or lettuce
+def create_and_configure_wrapper(context):
+    """Create and configure driver wrapper in behave tests
 
-    :param context_or_world: behave context or lettuce world
-    :param scenario: running scenario
-    :param no_driver: True if this is an api test and driver should not be started
-    """
-    # Initialize and connect driver wrapper
-    start_driver(context_or_world, no_driver)
-
-    # Add assert screenshot methods with scenario configuration
-    add_assert_screenshot_methods(context_or_world, scenario)
-
-    # Configure Jira properties
-    save_jira_conf()
-
-    context_or_world.logger.info("Running new scenario: %s", scenario.name)
-
-
-def create_and_configure_wrapper(context_or_world):
-    """Create and configure driver wrapper in behave or lettuce tests
-
-    :param context_or_world: behave context or lettuce world
+    :param context: behave context
     """
     # Create default driver wrapper
-    context_or_world.driver_wrapper = DriverWrappersPool.get_default_wrapper()
-    context_or_world.utils = context_or_world.driver_wrapper.utils
+    context.driver_wrapper = DriverWrappersPool.get_default_wrapper()
+    context.utils = context.driver_wrapper.utils
 
     # Get behave userdata properties to override config properties
     try:
-        behave_properties = context_or_world.config.userdata
+        behave_properties = context.config.userdata
     except AttributeError:
         behave_properties = None
 
     # Configure wrapper
-    context_or_world.driver_wrapper.configure(context_or_world.config_files, behave_properties=behave_properties)
+    context.driver_wrapper.configure(context.config_files, behave_properties=behave_properties)
 
     # Copy config object
-    context_or_world.toolium_config = context_or_world.driver_wrapper.config
+    context.toolium_config = context.driver_wrapper.config
 
     # Configure logger
-    context_or_world.logger = logging.getLogger(__name__)
+    context.logger = logging.getLogger(__name__)
 
 
-def connect_wrapper(context_or_world):
-    """Connect driver in behave or lettuce tests
+def connect_wrapper(context):
+    """Connect driver in behave tests
 
-    :param context_or_world: behave context or lettuce world
+    :param context: behave context
     """
     # Create driver if it is not already created
-    if context_or_world.driver_wrapper.driver:
-        context_or_world.driver = context_or_world.driver_wrapper.driver
+    if context.driver_wrapper.driver:
+        context.driver = context.driver_wrapper.driver
     else:
-        context_or_world.driver = context_or_world.driver_wrapper.connect()
+        context.driver = context.driver_wrapper.connect()
 
     # Copy app_strings object
-    context_or_world.app_strings = context_or_world.driver_wrapper.app_strings
+    context.app_strings = context.driver_wrapper.app_strings
 
 
-def add_assert_screenshot_methods(context_or_world, scenario):
-    """Add assert screenshot methods to behave or lettuce object
+def add_assert_screenshot_methods(context, scenario):
+    """Add assert screenshot methods to behave object
 
-    :param context_or_world: behave context or lettuce world
+    :param context: behave context
     :param scenario: running scenario
     """
     file_suffix = scenario.name
@@ -212,8 +202,8 @@ def add_assert_screenshot_methods(context_or_world, scenario):
         VisualTest(self.driver_wrapper, force).assert_screenshot(self.web_element, filename, file_suffix, threshold,
                                                                  exclude_elements)
 
-    context_or_world.assert_screenshot = assert_screenshot
-    context_or_world.assert_full_screenshot = assert_full_screenshot
+    context.assert_screenshot = assert_screenshot
+    context.assert_full_screenshot = assert_full_screenshot
     PageElement.assert_screenshot = assert_screenshot_page_element
 
 
@@ -223,31 +213,21 @@ def after_scenario(context, scenario):
     :param context: behave context
     :param scenario: running scenario
     """
-    bdd_common_after_scenario(context, scenario, scenario.status)
-
-
-def bdd_common_after_scenario(context_or_world, scenario, status):
-    """Clean method that will be executed after each scenario in behave or lettuce
-
-    :param context_or_world: behave context or lettuce world
-    :param scenario: running scenario
-    :param status: scenario status (passed, failed or skipped)
-    """
-    if status == 'skipped':
+    if scenario.status == 'skipped':
         return
-    elif status == 'passed':
+    elif scenario.status == 'passed':
         test_status = 'Pass'
         test_comment = None
-        context_or_world.logger.info("The scenario '%s' has passed", scenario.name)
+        context.logger.info("The scenario '%s' has passed", scenario.name)
     else:
         test_status = 'Fail'
         test_comment = "The scenario '%s' has failed" % scenario.name
-        context_or_world.logger.error("The scenario '%s' has failed", scenario.name)
-        context_or_world.global_status['test_passed'] = False
+        context.logger.error("The scenario '%s' has failed", scenario.name)
+        context.global_status['test_passed'] = False
 
     # Close drivers
-    DriverWrappersPool.close_drivers(scope='function', test_name=scenario.name, test_passed=status == 'passed',
-                                     context=context_or_world)
+    DriverWrappersPool.close_drivers(scope='function', test_name=scenario.name, test_passed=scenario.status == 'passed',
+                                     context=context)
 
     # Save test status to be updated later
     add_jira_status(get_jira_key_from_scenario(scenario), test_status, test_comment)
@@ -289,17 +269,9 @@ def after_all(context):
 
     :param context: behave context
     """
-    bdd_common_after_all(context)
-
-
-def bdd_common_after_all(context_or_world):
-    """Common after all method in behave or lettuce
-
-    :param context_or_world: behave context or lettuce world
-    """
     # Close drivers
     DriverWrappersPool.close_drivers(scope='session', test_name='multiple_tests',
-                                     test_passed=context_or_world.global_status['test_passed'])
+                                     test_passed=context.global_status['test_passed'])
 
     # Update tests status in Jira
     change_all_jira_status()